@@ -1,35 +1,56 @@
-import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-import { FormsModule } from '@angular/forms';
 import { RouterModule } from '@angular/router';
-import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
+import { HttpClientModule, HTTP_INTERCEPTORS, HttpClient } from '@angular/common/http';
+import { BrowserModule } from '@angular/platform-browser';
 
-import { ModalModule } from 'ngx-bootstrap/modal';
 
 import { AppComponent } from './app.component';
 import { NavMenuComponent } from './nav-menu/nav-menu.component';
 import { HomeComponent } from './home/home.component';
-import { CounterComponent } from './counter/counter.component';
 import { AuthorizeInterceptor } from 'src/api-authorization/authorize.interceptor';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { TranslateLoader, TranslateModule } from '@ngx-translate/core';
+import { TranslateHttpLoader } from '@ngx-translate/http-loader';
+import { MobileAccountComponent } from './mobile-account/mobile-account.component';
+import { MobileItemsComponent } from './mobile-items/mobile-items.component';
+import { MobileLoreComponent } from './mobile-lore/mobile-lore.component';
+import { MobileSpellsComponent } from './mobile-spells/mobile-spells.component';
+import { SwipeGestureDirective } from '../directives/swipe-gesture/swipe-gesture.directive';
+import { PcSliderComponent } from './pc-slider/pc-slider.component';
+
+export function createTranslateLoader(http: HttpClient){
+  return new TranslateHttpLoader(http, '/assets/i18n/', '.json');
+}
 
 @NgModule({
   declarations: [
     AppComponent,
     NavMenuComponent,
     HomeComponent,
-    CounterComponent
+    MobileLoreComponent,
+    MobileItemsComponent,
+    MobileSpellsComponent,
+    MobileAccountComponent,
+    PcSliderComponent
   ],
   imports: [
-    BrowserModule.withServerTransition({ appId: 'ng-cli-universal' }),
+    BrowserModule,
     HttpClientModule,
-    FormsModule,
+    TranslateModule.forRoot({
+      defaultLanguage: 'en',
+      useDefaultLang: true,
+      loader: {
+        provide: TranslateLoader,
+        useFactory: createTranslateLoader,
+        deps: [HttpClient]
+      }
+    }),
     RouterModule.forRoot([
-      { path: '', component: HomeComponent, pathMatch: 'full' },
-      { path: 'counter', component: CounterComponent }
+      { path: 'items', component: MobileItemsComponent },
+      { path: 'lore', component: MobileLoreComponent },
+      { path: 'spells', component: MobileSpellsComponent },
+      { path: 'account', component: MobileAccountComponent }
     ]),
-    BrowserAnimationsModule,
-    ModalModule.forRoot()
+    SwipeGestureDirective
   ],
   providers: [
     { provide: HTTP_INTERCEPTORS, useClass: AuthorizeInterceptor, multi: true }
